+++ samples/insert_with_continue_on_error_sample.py
@@ -1,0 +1,44 @@
+'''
+	Copyright (c) 2022 Skyflow, Inc.
+'''
+from skyflow.errors import SkyflowError
+from skyflow.service_account import generate_bearer_token, is_expired
+from skyflow.vault import Client, InsertOptions, Configuration
+
+# cache token for reuse
+bearerToken = ''
+
+def token_provider():
+    global bearerToken
+    if is_expired(bearerToken):
+        bearerToken, _ = generate_bearer_token('<YOUR_CREDENTIALS_FILE_PATH>')
+    return bearerToken
+
+
+try:
+    config = Configuration(
+       '<YOUR_VAULT_ID>', '<YOUR_VAULT_URL>', token_provider)
+    client = Client(config)
+
+    options = InsertOptions(tokens=True, continueOnError=True)
+
+    data = {
+        "records": [
+            {
+                "table": "<TABLE_NAME>",
+                "fields": {
+                    "<FIELDNAME>": "<VALUE>"
+                }
+            },
+            {
+                "table": "<TABLE_NAME>",
+                "fields": {
+                    "<FIELDNAME>": "<VALUE>"
+                }
+            }
+        ]
+    }
+    response = client.insert(data, options=options)
+    print('Response:', response)
+except SkyflowError as e:
+    print('Error Occurred:', e)